kg_pipeline/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
coloredlogs==15.0.1
cryptography==46.0.5
cssselect==1.4.0
distro==1.9.0
docopt==0.6.2
eutils==0.6.1
frozenlist==1.8.0
google-auth==2.48.0
google-genai==1.64.0
h11==0.16.0
habanero==2.3.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.11
Levenshtein==0.27.3
lxml==6.0.2
lxml_html_clean==0.4.3
metapub==0.6.4
multidict==6.7.1
neo4j==6.1.0
packaging==26.0
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-Levenshtein==0.27.3
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.32.5
rsa==4.9.1
setuptools==82.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
tabulate==0.9.0
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
Unidecode==1.4.0
urllib3==2.6.3
websockets==15.0.1
yarl==1.22.0
